+++ src/utils.js
@@ -6,7 +6,7 @@ const _           = require('lodash')
 const yaml        = require('js-yaml')
 const shell       = require('shelljs')
 // const spawnSync   = require('spawn-sync')
-const oneLine     = require('common-tags/lib/oneLine')
+// const oneLine     = require('common-tags/lib/oneLine')
 // const stripIndent = require('common-tags/lib/stripIndent')
 const scrolex     = require('scrolex').persistOpts({
   announce             : true,