+++ src/utils.test.js
@@ -1,9 +1,9 @@
-const utils         = require('./utils')
+// const utils         = require('./utils')
 // const debug      = require('depurar')('sut')
-const sut           = utils
-const cacheDir      = 'CACHEDIR'
-const projectDir    = 'PROJECTDIR'
-const lanyonVersion = 'LANYONVERSION'
+// const sut           = utils
+// const cacheDir      = 'CACHEDIR'
+// const projectDir    = 'PROJECTDIR'
+// const lanyonVersion = 'LANYONVERSION'
 
 describe('utils', () => {
 })